Enhance defcustom definitions
While there is a `list' type in `defcustom', it is meant for a very specific list. '(list string) means that you want a list with exactly one string. '(list string function) means that the customizable variable should take two elements: a string and a function.

However, we want arbitrary many strings, that's what `repeat' is for. Note that we could probably use `file' instead of `string', but the auto-completion does not recognize the `dotfiles-folder' as search-folder. Oh well.

For directories there is the special type 'directory'.