feat: Add ability to automatically org-babel based on file category
Using Emacs is a tool for productivity which performs functions with minimal
effort. Therefore, I have added a functionality that makes org-babel run in
`org-mode` based on the file level property which the user chooses. This will
help to give a productivity boost to the user.

If applied, this commit will:
  - Run `orb-babel` automatically when the file level `CATEGORY` property is present
    and have a value of `Configuration`